Inventory Refactor, Episode 3, Part 6 (finale)
IT FUCKIN' WORKS AND SAVES, LETS GOOOO
inventory now is accessible from the Hangar button in main menu, it saves both the mech's inventory and your stash inventory to Profile, and it all pretty much works. Next up is actually making it look nice, but for now the functionality is there. Merge happening right after this!